Provide single-file import of relevant manifest code for integrators
In addition to providing down-level implementation that can compile as-is in NS2 (typical for roslyn analyzers), we provide it all in a single file SponsorLink.cs instead, so it's easier to "clone" (or `dotnet file add/sync`).
