--- src/Commands/Constants.cs
@@ -1,34 +1,0 @@
-﻿namespace Devlooped.SponsorLink;
-
-/// <summary>
-/// Environment variable names used by the SponsorLink CLI.
-/// </summary>
-public class Constants
-{
-    /// <summary>
-    /// Last used access token to invoke SponsorLink backend APIs.
-    /// </summary>
-    public const string AccessTokenVariable = "SPONSORLINK_TOKEN";
-
-    /// <summary>
-    /// Random GUID used for salting hashes. Unique per installation, can be 
-    /// regenerated by deleting the environment variable.
-    /// </summary>
-    public const string InstallationIdVariable = "SPONSORLINK_INSTALLATION";
-
-    /// <summary>
-    /// JWT token containing the hashed claims that represent active 
-    /// sponsorships.
-    /// </summary>
-    public const string ManifestVariable = "SPONSORLINK_MANIFEST";
-
-    /// <summary>
-    /// Default audience for the manifest token.
-    /// </summary>
-    public const string ManifestAudience = "SponsorLink";
-
-    /// <summary>
-    /// Default issuer for the manifest token.
-    /// </summary>
-    public const string ManifestIssuer = "Devlooped";
-}
