+++ src/Commands/SponsorLink.cs
@@ -12,9 +12,83 @@ using Microsoft.IdentityModel.Tokens;
 namespace Devlooped.SponsorLink;
 
 /// <summary>
+/// Environment variable names used by the SponsorLink CLI.
+/// </summary>
+public class Constants
+{
+    /// <summary>
+    /// Last used access token to invoke SponsorLink backend APIs.
+    /// </summary>
+    public const string AccessTokenVariable = "SPONSORLINK_TOKEN";
+
+    /// <summary>
+    /// Random GUID used for salting hashes. Unique per installation, can be 
+    /// regenerated by deleting the environment variable.
+    /// </summary>
+    public const string InstallationIdVariable = "SPONSORLINK_INSTALLATION";
+
+    /// <summary>
+    /// JWT token containing the hashed claims that represent active 
+    /// sponsorships.
+    /// </summary>
+    public const string ManifestVariable = "SPONSORLINK_MANIFEST";
+
+    /// <summary>
+    /// Default audience for the manifest token.
+    /// </summary>
+    public const string ManifestAudience = "SponsorLink";
+
+    /// <summary>
+    /// Default issuer for the manifest token.
+    /// </summary>
+    public const string ManifestIssuer = "Devlooped";
+}
+
+/// <summary>
+/// Provides simplified access to the key environment variables.
+/// </summary>
+public static partial class Variables
+{
+    /// <summary>
+    /// Last used access token to invoke SponsorLink backend APIs.
+    /// </summary>
+    public static string? AccessToken
+    {
+        get => Environment.GetEnvironmentVariable(Constants.AccessTokenVariable, EnvironmentVariableTarget.User) is string token ?
+            !string.IsNullOrEmpty(token) ?
+            token : null : null;
+        set => Environment.SetEnvironmentVariable(Constants.AccessTokenVariable, value, EnvironmentVariableTarget.User);
+    }
+
+    /// <summary>
+    /// Random GUID used for salting hashes. Unique per installation, can be 
+    /// regenerated by deleting the environment variable.
+    /// </summary>
+    public static string? InstallationId
+    {
+        get => Environment.GetEnvironmentVariable(Constants.InstallationIdVariable, EnvironmentVariableTarget.User) is string id ?
+            !string.IsNullOrEmpty(id) ?
+            id : null : null;
+        set => Environment.SetEnvironmentVariable(Constants.InstallationIdVariable, value, EnvironmentVariableTarget.User);
+    }
+
+    /// <summary>
+    /// Gets the JWT token containing the hashed claims that represent 
+    /// active sponsorships.
+    /// </summary>
+    public static string? Manifest
+    {
+        get => Environment.GetEnvironmentVariable(Constants.ManifestVariable, EnvironmentVariableTarget.User) is string manifest ?
+            !string.IsNullOrEmpty(manifest) ?
+            manifest : null : null;
+        set => Environment.SetEnvironmentVariable(Constants.ManifestVariable, value, EnvironmentVariableTarget.User);
+    }
+}
+
+/// <summary>
 /// Represents a manifest of sponsorship claims.
 /// </summary>
-public partial class Manifest
+public class Manifest
 {
     static readonly SHA256 sha = SHA256.Create();
     readonly string salt;
